Add user provider and check username endpoint; refactor user module and social auth service
