+++ backend/src/auth/services/social.auth.service.ts
@@ -1,7 +1,6 @@
 import { BadRequestException, Injectable, InternalServerErrorException } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
 import { SessionService } from 'src/session/session.service';
-import { UserService } from 'src/user/user.service';
 import { CryptoService } from '../providers/crypto.service';
 import { DeviceDto } from '../dtos/device.data.dto';
 import { ErrorMessages } from 'src/common/enums/error-messages.enum';
@@ -9,6 +8,7 @@ import { GoogleProfile } from '../../common/types/auth';
 import { User } from 'src/user/entities/user.entity';
 import { RandomHelper } from 'src/common/utils/random.helper';
 import { AuthEnum } from 'src/common/enums/auth.enum';
+import { UserProvider } from 'src/user/providers/user.provider';
 
 
 @Injectable()
@@ -16,26 +16,26 @@ export class SocialAuthService {
   constructor(
     private readonly jwtService: JwtService,
     private readonly sessionService: SessionService,
-    private readonly userService: UserService,
+    private readonly userProvider: UserProvider,
     private readonly cryptoService: CryptoService,
   ) { }
 
 
   private async createUser(profile: GoogleProfile): Promise<User> {
     // Check if user with email, username or phone number already exists
-    const userWithEmail = await this.userService.findByEmail(profile.email);
+    const userWithEmail = await this.userProvider.findByEmail(profile.email);
     if (userWithEmail) {
       throw new BadRequestException(ErrorMessages.EMAIL_ALREADY_EXISTS);
     }
 
     if (profile.phoneNumber) {
-      const userWithPhoneNumber = await this.userService.findByPhoneNumber(profile.phoneNumber);
+      const userWithPhoneNumber = await this.userProvider.findByPhoneNumber(profile.phoneNumber);
       if (userWithPhoneNumber) {
         throw new BadRequestException(ErrorMessages.PHONE_NUMBER_ALREADY_EXISTS);
       }
     }
 
-    const userWithUsername = await this.userService.findByUsername(profile.displayName);
+    const userWithUsername = await this.userProvider.findByUsername(profile.displayName);
     if (userWithUsername) {
       const randomString = RandomHelper.generateRandomString(5);
       const randomNumber = RandomHelper.generateRandomNumber(3);
@@ -43,7 +43,7 @@ export class SocialAuthService {
     }
 
     // Create user
-    const newUser = await this.userService.createGoogleUser({
+    const newUser = await this.userProvider.createGoogleUser({
       email: profile.email,
       username: profile.displayName,
       phoneNumber: profile.phoneNumber,
@@ -63,7 +63,7 @@ export class SocialAuthService {
 
 
   async createOrLogin(profile: GoogleProfile, deviceId: string) {
-    const user = await this.userService.findByGoogleId(profile.id) || await this.createUser(profile);
+    const user = await this.userProvider.findByGoogleId(profile.id) || await this.createUser(profile);
     const device = await this.sessionService.getDeviceInfo(deviceId);
 
     if (!device) {
