+++ backend/src/user/user.controller.ts
@@ -1,14 +1,15 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import { Controller, Get, HttpCode, HttpStatus, Query } from '@nestjs/common';
 import { UserService } from './user.service';
-import { CreateUserDto } from './dto/create-user.dto';
-import { UpdateUserDto } from './dto/update-user.dto';
+import { UserEndpoints } from 'src/common/enums/endpoints.enum';
 
 @Controller('user')
 export class UserController {
   constructor(private readonly userService: UserService) {}
 
-  @Post()
-  create(@Body() createUserDto: CreateUserDto) {
-    return this.userService.create(createUserDto);
+  @Get(UserEndpoints.CHECK_USERNAME)
+  @HttpCode(HttpStatus.OK)
+  async checkAvailability(@Query("username") username: string){
+    return this.userService.checkUsernameAvailability(username);
   }
+
 }
