+++ backend/src/user/user.module.ts
@@ -4,11 +4,12 @@ import { UserController } from './user.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from './entities/user.entity';
 import { CacheModule } from 'src/cache/cache.module';
+import { UserProvider } from './providers/user.provider';
 
 @Module({
   controllers: [UserController],
-  providers: [UserService],
+  providers: [UserService, UserProvider],
   imports: [TypeOrmModule.forFeature([User]), CacheModule],
-  exports: [UserService],
+  exports: [UserProvider],
 })
 export class UserModule {}
